GetALLUSers api working
