+++ go.mod
@@ -2,7 +2,11 @@ module api
 
 go 1.22.0
 
-require github.com/gofiber/fiber/v2 v2.52.2
+require (
+	github.com/doug-martin/goqu/v9 v9.19.0
+	github.com/gofiber/fiber/v2 v2.52.2
+	github.com/lib/pq v1.10.9
+)
 
 require (
 	github.com/andybalholm/brotli v1.0.5 // indirect
